CELIX-219: Update framework shutdown to clean up bundles that are still installed.
During shutdown of the framework, the shutdown thread has to close stopped bundles as well, this triggers the resolved to remove the module from the active list.
During destruction of the framework all bundles in the bundle list have to be cleaned up. This now includes closing the libraries opened during installation.
When modules are removed from the resolver, and the list is empty, the lists are now destroyed.

Fixed a missing dependency when building the examples
Added NULL initialiser to the hash_map.

git-svn-id: https://svn.apache.org/repos/asf/celix/trunk@1658959 13f79535-47bb-0310-9956-ffa450edef68