CELIX-215: Added curl init/cleanup to the launcher
To be able to use curl properly it has to be initialised once while no other threads are running. With Celix this cannot be done in a bundle, so to prevent problems during runtime, the launcher now calls init/cleanup.
While this is not yet conditional, it could definitely be made an option. Environments that do not use Curl probably don't want a launcher that depends on it.

git-svn-id: https://svn.apache.org/repos/asf/celix/trunk@1658960 13f79535-47bb-0310-9956-ffa450edef68