CELIX-389: Refactors pubsub.
Moves sources from celix-pubsub/pubsub to pubsub and  celix-pubsub/cmake to cmake. Renames the build option for ZMW to BUILD_PUBSUB_PSA_ZMQ, so that in the gui the options are next to each other. Uses etcdlib as static lib instead of shared lib. Adds the subdirectory pubsub to the root cmake file.