CELIX-452: Changes the behaviour of the serviceTracker_close, so that removing the service listener is done on a separate thread.
This is needed because the previous commits added sync behaviour to the serviceChanged handling. As result a serviceChange event will ensure that removeServiceListener call will only returned if it is not used anymore. This can introduces a deadlock when removing a service listener from a service change event callback. This is specifically the case with service trackers, because they remove a service listeners in the sync close call.

The removing of the service listener is now done on a detached thread. During bundle stop and framework stop a sync call is used to ensure all detached "service tracker shutdown" threads are done. Because the current api support stopping and restarted a service tracker, some considerable refactoring was needed in the general service tracker internals.