Updates the HTTP Admin config properties and adds resource urls info to the http_admin_info service.
The HTTP Admin properties did not contain any prefixes.
This is the case for many bundle config properties, but for config property we try to have an explicit prefix.
For HTTP Admin this is now CELIX_HTTP_ADMIN.

Also adds support for configured the num of threads used in civetweb (now default 4).