Refactors the handling of service listeners. This is moved to the service registry and the service listener is now retroactively creating events.
This is done because there was a race condition in the service trackers when adding service listener and getting a list of already registered services.
This was done with two calls and as result introduced a race condition.
Because the service listeners are changed in behaviour (now retroactively), the direct use if also  deprecated; use the service tracker instead.