Disables the use of service tracker shutdown threads and updates the logging tests to increase test coverage.
Closing a service tracker was done on a separate shutdown thread.
This was expensive and needed because of locking issue in the framework for service listeners.
Service listener callbacks are now called outsize of a lock and as result the shutdown threads
should not be needed anymore.
For now, the old behaviour is still available with a compiler flag
(CELIX_SERVICE_TRACKER_USE_SHUTDOWN_THREAD)