Bugfix/zmq wrong sender connections (#264)
* Fixes an issue in endpoint endpoint discovery.

* Changes handling of scopeless and default scoped topic sendes/receivers.

In the udpated setup a scopeless (scope==NULL) and default scoped (scope="default") will have their own topic sender and receiver, but will connect to each other through discovery. This is done so that requesting a scopeless publisher will result in a service without a scope property and requesting a default scoped publisher will result in a publisher with a "default" property.