Add README.md to links and fix the OSGI links
The README.md is needed since Hugo (the static website generator Celix uses) otherwise will create a link to a directory, where it searches for an index.html file instead of a README.html. Having the README.md in the link will ensure that both GitHub and Hugo handle the link correctly.